# src/main/resources/public/api-contract.yml
openapi: 3.0.1
info:
  title: Votação API
  description: |-
    Para obter mais informações sobre o desenvolvimento

    ## Overview

    Api CoopVotes

    Esta api está agnostica de clientes deste que uso protocolos, **HTTP**.

    ## Começando

    Esta documentação ira auxiliar em novas implementações de __clients sides__, dando uma base
    de como manipular os recursos que a API provem.

    ### HTTP Verbs

    Esta Api foi construída a seguir os padrões RESTFull, tentando aderir o mais próximo
    possíveis das convenções uso dos verbos HTTP.


    | Verbo     | Utilização                                      |
    | :-------: | :---------------------------------------------- |
    | `GET`     | Usado para recuperar recursos                   |
    | `POST`    | Usado para criar um recurso                     |
    | `PUT`     | Usado para atualizar um recurso                 |
    | `PATCH`   | Usado para atualizar um parcialmente um recurso |
    | `DELETE`  | Usado para remover um recurso existente         |

    ### HTTP status codes

    Esta Api foi construída seguindo os padrões RESTFull, tentando aderir o mais próximo
    possíveis as suas convenções em seu uso dos códigos de Status HTTP.

    | Status code                    | Utilização                                                                                                                          |
...

    # Header
...
       aqui dispostos.

    ## Requisições
    Headers para enviar requisições.

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /vote:
    post:
      summary: Include uma voto a uma pauta especifica.
      operationId: includeVote
      tags:
        - Vote
      description: Adiciona um voto
      requestBody:
        required: true
        description: Valor do voto.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputIncludeVote'
      responses:
        '202':
          description: Voto foi aceito com sucesso.
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    InputIncludeVote:
      type: object
      required:
        - idStave
        - legalDocumentNumberAssociate
        - value
      properties:
        idStave:
          type: integer
          format: int64
          description: Identificador a Pauta que será incluida o voto
          example: 10000
        legalDocumentNumberAssociate:
          type: string
          pattern: "([0-9]{3}[.]?[0-9]{3}[.]?[0-9]{3}-[0-9]{2})|([0-9]{11})"
          minLength: 11
          maxLength: 14
          description: Cpf do associado que está votando. Api aceita CPF com ou sem pontuação
          example: 111.111.111-11
        value:
          type: boolean
          nullable: false
          description: |-
             Valor do voto,
              * `True` para sim,
              * `Falso` para não
          example: true


    # erros
    ErrorBase:
      type: object
      properties:
        statusCode:
          type: integer
          example: 0
          description: Code de status http
        message:
          type: string
          example: Um erro ocorreu.
          description: Mensagem da ocorrencia do erro.
        path:
          type: string
          example: /path
          description: uri da ocorrencia do erro
    Error:
      allOf:
        - type: object
        - $ref: '#/components/schemas/ErrorBase'
        - properties:
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetails'
    ErrorDetails:
      type: object
      properties:
        field:
          type: string
          example: type.field
          description: nome do campo que contém um erro
        descriptionError:
          type: string
          example: Campo inválido
          description: descreve o erro encontrado no campo.
  headers:
    Version-Api:
      description: Define a versão da api, da qual respondeu a solicitação.
      schema:
        type: string
        default: '1.0.0'
      example: '0.0.1'
  responses:
...
    BadRequest:
      description: >
        Requisição está invalida, o corpo da mensagem está mal formatada, por exemplo, campos faltando ou
        campos estão inválidos (formato inválido).
          No corpo da resposta encontrar uma lista de detalhes que dará informações pertinentes para corrigir o erro.
        Para maiores informações consulte a [rfc-7231](https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.1).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Version-Api:
          $ref: '#/components/headers/Version-Api'
...
    InternalServerError:
      description: >
        Error inesperado em alguma parte do processo, neste caso a api não conseguiu se recuperar
        do error.


        Caso isto ocorra, por favor, notifique o momento do erro, crie uma issue no __repositório__ da aplicação explicando
        os detalhes de como aconteceu o error.


        Para maiores informações consulte a [rfc-7231](https://datatracker.ietf.org/doc/html/rfc7231#section-6.6.1).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBase'